refactor: move url methods to utils/_url.py
