+++ cognite/client/_api/templates.py
@@ -23,8 +23,8 @@ from cognite.client.data_classes.templates import (
     ViewResolveList,
     ViewWrite,
 )
-from cognite.client.utils._auxiliary import interpolate_and_url_encode
 from cognite.client.utils._identifier import IdentifierSequence
+from cognite.client.utils._url import interpolate_and_url_encode
 from cognite.client.utils.useful_types import SequenceNotStr
 
 if TYPE_CHECKING:
